+++ modules/client/src/addresslistVoting/internal/constants.ts
@@ -1,7 +1,7 @@
 import {
   AddresslistVoting__factory,
   MajorityVotingBase__factory,
-} from "@aragon/core-contracts-ethers";
+} from "@aragon/osx-ethers";
 
 export const AVAILABLE_FUNCTION_SIGNATURES: string[] = [
   MajorityVotingBase__factory.createInterface().getFunction(