+++ modules/client/src/client-common/constants.ts
@@ -1,5 +1,5 @@
-import { activeContractsList } from "@aragon/core-contracts-ethers";
-import { SupportedNetworks, NetworkDeployment } from "./interfaces/common";
+import { activeContractsList } from "@aragon/osx-ethers";
+import { NetworkDeployment, SupportedNetworks } from "./interfaces/common";
 import { ProposalMetadata } from "./interfaces/plugin";
 
 export const UNSUPPORTED_PROPOSAL_METADATA_LINK: ProposalMetadata = {
@@ -17,18 +17,19 @@ export const UNAVAILABLE_PROPOSAL_METADATA: ProposalMetadata = {
 
 export const LIVE_CONTRACTS: { [K in SupportedNetworks]: NetworkDeployment } = {
   mainnet: {
-    // TODO change to mainnet
-    daoFactory: activeContractsList.goerli.DAOFactory,
-    multisigRepo: "multisig.plugin.dao.eth",
-    adminRepo: "admin.plugin.dao.eth",
-    addresslistVotingRepo: "address-list-voting.plugin.dao.eth",
-    tokenVotingRepo: "token-voting.plugin.dao.eth",
+    daoFactory: activeContractsList.mainnet.DAOFactory,
+    multisigRepo: activeContractsList.mainnet["multisig-repo"],
+    adminRepo: activeContractsList.mainnet["admin-repo"],
+    addresslistVotingRepo:
+      activeContractsList.mainnet["address-list-voting-repo"],
+    tokenVotingRepo: activeContractsList.mainnet["token-voting-repo"],
   },
   goerli: {
     daoFactory: activeContractsList.goerli.DAOFactory,
     multisigRepo: activeContractsList.goerli["multisig-repo"],
     adminRepo: activeContractsList.goerli["admin-repo"],
-    addresslistVotingRepo: activeContractsList.goerli["address-list-voting-repo"],
+    addresslistVotingRepo:
+      activeContractsList.goerli["address-list-voting-repo"],
     tokenVotingRepo: activeContractsList.goerli["token-voting-repo"],
   },
 };