+++ modules/client/src/internal/client/estimation.ts
@@ -3,7 +3,7 @@ import {
   DAOFactory,
   DAOFactory__factory,
   PluginRepo__factory,
-} from "@aragon/core-contracts-ethers";
+} from "@aragon/osx-ethers";
 import {
   InvalidAddressOrEnsError,
   NoProviderError,
@@ -76,7 +76,7 @@ export class ClientEstimation extends ClientCore implements IClientEstimation {
 
     const gasEstimation = await daoInstance.estimateGas.createDao(
       {
-        name: params.ensSubdomain,
+        subdomain: params.ensSubdomain,
         metadata: toUtf8Bytes(params.metadataUri),
         daoURI: params.daoUri || "",
         trustedForwarder: params.trustedForwarder || AddressZero,